Wizard spells (#11307)
* Wizard spells

* fix

* Fixes dark matter  explossion Being offset

* Namespace aand stuff

* Cleanup

* fix

* Fixed tests

* fix

+++ UnityProject/Assets/Scripts/US13/Objects/InternalBattery.cs
@@ -68,10 +68,10 @@ namespace US13.Objects
 		}
 
 
-		public float MaxCharge {
+		public int MaxCharge {
 			get
 			{
-				float maxwatts = 0;
+				int maxwatts = 0;
 				foreach (Battery battery in batteries.Values)
 				{
 					maxwatts += battery.MaxWatts;
